Reshuffled document type tests
