+++ tests/config/test_validate_document_type_config.py
@@ -4,7 +4,6 @@ import pytest
 
 from briefcase.config import validate_document_type_config
 from briefcase.exceptions import BriefcaseConfigError
-from briefcase.platforms.macOS import utils
 
 
 @pytest.fixture
@@ -214,30 +213,6 @@ def test_document_type_macOS_config_with_unknown_mimetype(valid_document):
         "public.data",
         "public.content",
     ]
-
-
-@pytest.mark.skipif(sys.platform != "darwin", reason="Test runs only on macOS")
-def test_is_uti_core_type():
-    """Check if a UTI is a core type."""
-    assert utils.is_uti_core_type(None) is False
-    assert utils.is_uti_core_type("public.data") is True
-    assert utils.is_uti_core_type("public.content") is True
-    assert utils.is_uti_core_type("com.adobe.pdf") is True
-    assert utils.is_uti_core_type("com.unknown.data") is False
-
-
-@pytest.mark.skipif(sys.platform != "darwin", reason="Test runs only on macOS")
-def test_mime_type_to_uti_with_nonexisting_coretypes_file(monkeypatch):
-    """Test that mime_type_to_UTI returns None if the coretypes file doesn't exist."""
-    monkeypatch.setattr(utils, "CORETYPES_PATH", "/does/not/exist")
-    assert utils.mime_type_to_uti("application/pdf") is None
-
-
-@pytest.mark.skipif(sys.platform != "darwin", reason="Test runs only on macOS")
-def test_is_uti_core_type_with_nonexisting_coretypes_file(monkeypatch):
-    """Test that mime_type_to_UTI returns None if the coretypes file doesn't exist."""
-    monkeypatch.setattr(utils, "CORETYPES_PATH", "/does/not/exist")
-    assert utils.is_uti_core_type("com.adobe.pdf") is False
 
 
 @pytest.mark.skipif(sys.platform != "darwin", reason="Test runs only on macOS")
